Data.Money.Internal

Description

This is an internal module. Import Data.Money instead.

Dense monetary values

data Dense currency

Dense represents a dense monetary value for currency (usually a ISO-4217 currency code, but not necessarily) as a rational number.

While monetary values associated with a particular currency are discrete, you can still treat monetary values as dense while operating on them. For example, the half of USD 3.41 is USD 1.705, which is not an amount that can't be represented as a number of USD cents (the smallest unit that can represent USD amounts). Nevertheless, if you eventually multiply USD 1.705 by 4, for example, you end up with USD 6.82, which is again a value representable as USD cents. In other words, Dense monetary values allow us to perform precise calculations deferring the conversion to a Discrete monetary values as much as posible. Once you are ready to aproximate a Dense value to a Discrete value you can use one of round, floor, ceiling or truncate. Otherwise, using toRational you can obtain a precise Rational representation.

Construct Dense monetary values using dense, or fromInteger / fromIntegral if that suffices.

WARNING if you want to treat a dense monetary value as a Real number (for example, to take the square root of that monetary value), then you are on your own. We can only guarantee lossless manipulation of rational values, so you will need to convert back and forth betwen the Rational representation for Dense and your (likely lossy) representation for Real numbers.

dense :: Rational -> Maybe (Dense currency)

Build a Dense monetary value from a Rational value.

For example, if you want to represent USD 12.52316, then you can use:

dense (125316 % 10000)

This function returns Nothing in case the given Rational is infinity or notANumber.

Discrete monetary values

type Discrete currency unit = Discrete' currency (Scale currency unit)

Discrete represents a discrete monetary value for a currency expresed as an integer amount of a particular unit. For example, with currency ~ "USD" and unit ~ "cent" you can represent United States Dollars to their full extent.

currency is usually a ISO-4217 currency code, but not necessarily.

Construct Discrete values using fromInteger.

For example, if you want to represent GBP 21.05, where the smallest represetable unit for a GBP (United Kingdom Pound) is the penny, and 100 pennies equal 1 GBP (i.e., Scale "GBP" ~ '(100, 1)), then you can use:

fromInteger 2105 :: Discrete "GBP" "penny"

Because 2015 / 100 == 20.15.

data Discrete' currency scale

Discrete' represents a discrete monetary value for a currency expresed as an amount of scale, which is a rational number expressed as (numerator, denominator).

You'll be using Discrete instead of Discrete' most of the time, which mentions the unit name (such as cent or centavo) instead of explicitely mentioning the unit scale.

fromDiscrete

Arguments

:: GoodScale scale

=> Discrete' currency scale

-> Dense currency

Convert currency Discrete monetary value into a Dense monetary value.

round

Arguments

:: GoodScale scale

=> Dense currency

-> (Discrete' currency scale, Maybe (Dense currency))

Round a Dense value x to the nearest value fully representable in its currency's unit Scale, which might be x itself.

If x is already fully representable in its currency's unit Scale, then the following holds:

round x == (x, Nothing)

Otherwise, if the nearest value to x that is fully representable in its currency's unit Scale is greater than x, then the following holds:

round == ceiling

Otherwise, the nearest value to x that is fully representable in its currency's unit Scale is smaller than x, and the following holds:

round == floor

Proof that round doesn't lose money:

x == case round x of
       (y, Nothing) -> y
       (y, Just z)  -> y + z

ceiling

Arguments

:: GoodScale scale

=> Dense currency

-> (Discrete' currency scale, Maybe (Dense currency))

Round a Dense value x to the nearest value fully representable in its currency's unit Scale which is greater than x or equal to x.

If x is already fully representable in its currency's unit Scale, then the following holds:

ceiling x == (x, Nothing)

Otherwise, if x is not representable in its currency's unit Scale, then the following holds:

ceiling x == (y, Just z)

x /= y

z < zero

Proof that ceiling doesn't lose money:

x == case ceiling x of
       (y, Nothing) -> y
       (y, Just z)  -> y + z

floor

Arguments

:: GoodScale scale

=> Dense currency

-> (Discrete' currency scale, Maybe (Dense currency))

Round a Dense value x to the nearest value fully representable in its currency's unit Scale which is smaller than x or equal to x.

If x is already fully representable in its currency's unit Scale, then the following holds:

floor x == (x, Nothing)

Otherwise, if x is not representable in its currency's unit Scale, then the following holds:

floor x == (y, Just z)

x /= y

z > zero

Proof that floor doesn't lose money:

x == case floor x of
       (y, Nothing) -> y
       (y, Just z)  -> y + z

truncate

Arguments

:: GoodScale scale

=> Dense currency

-> (Discrete' currency scale, Maybe (Dense currency))

Round a Dense value x to the nearest value between zero and x (inclusive) which is fully representable in its currency's unit Scale.

If x is already fully representable in its currency's unit Scale, then the following holds:

truncate x == (x, Nothing)

Otherwise, if x is positive, then the following holds:

truncate == floor

Otherwise, if x is negative, the following holds:

truncate == ceiling

Proof that truncate doesn't lose money:

x == case truncate x of
       (y, Nothing) -> y
       (y, Just z)  -> y + z

Currency scales

type family Scale (currency :: Symbol) (unit :: Symbol) :: (Nat, Nat)

Scale currency unit is a rational number (expressed as '(numerator, denominator)) indicating how many pieces of unit fit in currency.

currency is usually a ISO-4217 currency code, but not necessarily.

The Scale will determine how to convert a Dense value into a Discrete value and vice-versa.

For example, there are 100 USD cents in 1 USD, so the scale for this relationship is:

type instance Scale "USD" "cent" = '(100, 1)

As another example, there is 1 dollar in USD, so the scale for this relationship is:

type instance Scale "USD" "dollar" = '(1, 1)

When using Discrete values to represent money, it will be impossible to represent an amount of currency smaller than unit. So, if you decide to use Scale "USD" "dollar" as your scale, you will not be able to represent values such as USD 3.50 or USD 21.87, since they are not exact multiples of a dollar.

If there exists a cannonical smallest unit that can fully represent the currency, then an instance Scale currency currency exists.

type instance Scale "USD" "USD" = Scale "USD" "cent"

For some monetary values, such as precious metals, the smallest representable unit is not obvious, since you can continue to split the precious metal many times before it stops being a precious metal. Still, for practical purposes we can make a sane arbitrary choice of smallest unit. For example, the base unit for XAU (Gold) is the troy ounce, which is too big to be considered the smallest unit, but we can arbitrarily choose the milligrain as our smallest unit, which is about as heavy as a single grain of table salt and should be sufficiently precise for all monetary practical purposes. A /troy ounce equals 480000 milligrains/.

type instance Scale "XAG" "milligrain" = '(480000, 1)

You can use other units such as milligrams for measuring XAU, for example. However, since the amount of milligrams in a troy ounce (31103.477) is not integral, we need to use rational number to express it.

type instance Scale "XAU" "milligram" = '(31103477, 1000)

If you try to obtain the 'Scale of a currency without an obvious smallest representable unit, like XAU, you will get a compile error.

type GoodScale scale = (CmpNat 0 (Fst scale) ~ LT, CmpNat 0 (Snd scale) ~ LT, KnownNat (Fst scale), KnownNat (Snd scale))

Constraints to a scale (like the one returned by Scale currency unit) expected to always be satisfied. In particular, the scale is always guaranteed to be a positive rational number (infinity and notANumber are forbidden by GoodScale).

type family ErrScaleNonCanonical (currency :: Symbol) :: k where ...

A friendly TypeError to use for a currency that doesn't have a cannonical small unit.

Equations

ErrScaleNonCanonical c = TypeError ((Text c :<>: Text " is not a currency with a canonical smallest unit,") :$$: Text "be explicit about the currency unit you want to use.")

scale

Arguments

:: GoodScale scale

=> proxy scale

-> Rational

Term-level representation for the currency's unit Scale.

For example, the Scale for "USD" in "cent"s is 100/1.

The returned Rational is statically guaranteed to be a positive number, and to be different from both notANumber and infinity.

Currency exchange

data ExchangeRate src dst

Exchange rate for converting monetary values of currency src into monetary values of currency dst by multiplying for it.

For example, if in order to convert USD to GBP we have to multiply by 1.2345, then we can represent this situaion using:

exchangeRate (12345 % 10000) :: Maybe (ExchangeRate "USD" "GBP")

exchangeRate :: Rational -> Maybe (ExchangeRate src dst)

Safely construct an ExchangeRate from a Rational number.

For construction to succeed, this Rational must be greater than 0, different from infinity and different from notANumber.

fromExchangeRate :: ExchangeRate src dst -> Rational

Obtain a Rational representation of the ExchangeRate.

This Rational is statically guaranteed to be greater than 0, different from infinity and different from notANumber.

flipExchangeRate :: ExchangeRate a b -> ExchangeRate b a

Flip the direction of an ExchangeRate.

Identity law:

flipExchangeRate . flipExchangeRate   ==  id

exchange :: ExchangeRate src dst -> Dense src -> Dense dst

Apply the ExchangeRate to the given Dense src monetary value.

Identity law:

exchange (flipExchangeRate x) . exchange x  ==  id

Use the Identity law for reasoning about going back and forth between src and dst in order to manage any leftovers that might not be representable as a Discrete monetary value of src.

Serializable representations

data DenseRep

A monomorphic representation of Dense that is easier to serialize and deserialize than Dense in case you don't know the type indexes involved.

denseRep :: KnownSymbol currency => Dense currency -> DenseRep

Convert a Dense to a DenseRep for ease of serialization.

denseRepCurrency :: DenseRep -> String

Currency name.

denseRepAmount :: DenseRep -> Rational

Currency unit amount.

denseRepAmountNumerator :: DenseRep -> Integer

Currency unit amount numerator.

denseRepAmountDenominator :: DenseRep -> Integer

Currency unit amount denominator. Positive, non-zero.

mkDenseRep

Arguments

:: String
Currency.

-> Integer
Scale nominator.

-> Integer
Scale denominator (positive, non zero)

-> Maybe DenseRep

Internal. Build a DenseRep from raw values.

fromDenseRep

Arguments

:: KnownSymbol currency

=> DenseRep

-> Maybe (Dense currency)

Attempt to convert a DenseRep to a Dense, provided you know the target currency.

withDenseRep

Arguments

:: DenseRep

-> (forall currency. KnownSymbol currency => Dense currency -> r)

-> r

Convert a DenseRep to a Dense without knowing the target currency.

Notice that currency here can't leave its intended scope unless you can prove equality with some other type at the outer scope, but in that case you would be better off using fromDenseRep directly.

data DiscreteRep

A monomorphic representation of Discrete that is easier to serialize and deserialize than Discrete in case you don't know the type indexes involved.

discreteRep

Arguments

:: (KnownSymbol currency, GoodScale scale)

=> Discrete' currency scale

-> DiscreteRep

Convert a Discrete to a DiscreteRep for ease of serialization.

discreteRepCurrency :: DiscreteRep -> String

Currency name.

discreteRepScale :: DiscreteRep -> Rational

Positive, non-zero.

discreteRepScaleNumerator :: DiscreteRep -> Integer

Positive, non-zero.

discreteRepScaleDenominator :: DiscreteRep -> Integer

Positive, non-zero.

discreteRepAmount :: DiscreteRep -> Integer

Amount of currency unit.

mkDiscreteRep

Arguments

:: String
Currency name.

-> Integer
Scale numerator. Positive, non-zero.

-> Integer
Scale denominator. Positive, non-zero.

-> Integer
Amount of unit.

-> Maybe DiscreteRep

Internal. Build a DiscreteRep from raw values.

fromDiscreteRep

Arguments

:: (KnownSymbol currency, GoodScale scale)

=> DiscreteRep

-> Maybe (Discrete' currency scale)

Attempt to convert a DiscreteRep to a Discrete, provided you know the target currency and unit.

withDiscreteRep

Arguments

:: DiscreteRep

-> (forall currency scale. (KnownSymbol currency, GoodScale scale) => Discrete' currency scale -> r)

-> r

Convert a DiscreteRep to a Discrete without knowing the target currency and unit.

Notice that currency and unit here can't leave its intended scope unless you can prove equality with some other type at the outer scope, but in that case you would be better off using fromDiscreteRep directly.

Notice that you may need to add an explicit type to the result of this function in order to keep the compiler happy.

data ExchangeRateRep

A monomorphic representation of ExchangeRate that is easier to serialize and deserialize than ExchangeRate in case you don't know the type indexes involved.

exchangeRateRep

Arguments

:: (KnownSymbol src, KnownSymbol dst)

=> ExchangeRate src dst

-> ExchangeRateRep

Convert a ExchangeRate to a DiscreteRep for ease of serialization.

exchangeRateRepSrcCurrency :: ExchangeRateRep -> String

Source currency name.

exchangeRateRepDstCurrency :: ExchangeRateRep -> String

Destination currency name.

exchangeRateRepRate :: ExchangeRateRep -> Rational

Exchange rate. Positive, non-zero.

exchangeRateRepRateNumerator :: ExchangeRateRep -> Integer

Exchange rate numerator. Positive, non-zero.

exchangeRateRepRateDenominator :: ExchangeRateRep -> Integer

Exchange rate denominator. Positive, non-zero.

mkExchangeRateRep

Arguments

:: String
Source currency name.

-> String
Destination currency name.

-> Integer
Exchange rate numerator. Positive, non-zero.

-> Integer
Exchange rate denominator. Positive, non-zero.

-> Maybe ExchangeRateRep

Internal. Build a ExchangeRateRep from raw values.

fromExchangeRateRep

Arguments

:: (KnownSymbol src, KnownSymbol dst)

=> ExchangeRateRep

-> Maybe (ExchangeRate src dst)

Attempt to convert a ExchangeRateRep to a ExchangeRate, provided you know the target src and dst types.

withExchangeRateRep

Arguments

:: ExchangeRateRep

-> (forall src dst. (KnownSymbol src, KnownSymbol dst) => ExchangeRate src dst -> r)

-> r

Convert a ExchangeRateRep to a ExchangeRate without knowing the target currency and unit.

Notice that src and dst here can't leave its intended scope unless you can prove equality with some other type at the outer scope, but in that case you would be better off using fromExchangeRateRep directly.