Money

Description

Import this module qualified as follows:

import qualified Money

Note: This module exports support for many well-known currencies out-of-the-box, but you are not limited to the currencies mentioned here. You can simply create a new UnitScale instance, and voilà. If you want to add a new currency to the out-of-the-box offer, please request so in https://github.com/k0001/safe-money/issues and the authors will see to it.

This module offers plenty of documentation, but for a deep explanation of how all of the pieces fit together, please read https://ren.zone/articles/safe-money. Notice, however, that this library has changed a bit since that article was written. You can always see the change log to understand what has changed.

Also, keep in mind that useful instances for the many types defined by safe-money can be found in these other libraries:

safe-money-aeson: FromJSON and ToJSON instances (from the aeson library).

safe-money-cereal: Serialize instances (from the cereal library).

safe-money-serialise: Serialise instances (from the serialise library).

safe-money-store: Store instances (from the store library).

safe-money-xmlbf: FromXml and ToXml instances (from the xmlbf library).

Dense monetary values

data Dense (currency :: Symbol)

Dense represents a dense monetary value for currency (usually a ISO-4217 currency code, but not necessarily) as a rational number.

While monetary values associated with a particular currency are discrete (e.g., an exact number of coins and bills), you can still treat monetary values as dense while operating on them. For example, the half of USD 3.41 is USD 1.705, which is not an amount that can't be represented as a number of USD cents (the smallest unit that can represent USD amounts). Nevertheless, if you do manage to represent USD 1.705 somehow, and you eventually multiply USD 1.705 by 4 for example, then you end up with USD 6.82, which is again a value representable as USD cents. In other words, Dense monetary values allow us to perform precise calculations deferring the conversion to a Discrete monetary values as much as posible. Once you are ready to approximate a Dense value to a Discrete value you can use one discreteFromDense. Otherwise, using toRational you can obtain a precise Rational representation.

denseCurrency :: KnownSymbol currency => Dense currency -> Text

Dense currency identifier.

> denseCurrency (dense' 4 :: Dense "USD")
"USD"

dense :: Rational -> Maybe (Dense currency)

Build a Dense monetary value from a Rational value.

For example, if you want to represent USD 12.52316, then you can use:

dense (125316 % 10000)

Notice that dense returns Nothing in case the given Rational's denominator is zero, which although unlikely, it is possible if the Rational was unsafely constructed. When dealing with hardcoded or trusted Rational values, you can use dense' instead of dense which unsafely constructs a Dense.

dense' :: Rational -> Dense currency

Unsafely build a Dense monetary value from a Rational value. Contrary to dense, this function *crashes* if the given Rational has zero as a denominator, which is something very unlikely to happen unless the given Rational was itself unsafely constructed. Other than that, dense and dense' behave the same.

Prefer to use dense when dealing with Rational inputs from untrusted sources.

denominator x /= 0
  ⇒ dense x == Just (dense' x)

denominator x == 0
  ⇒ undefined == dense' x

denseFromDiscrete

Arguments

:: GoodScale scale

=> Discrete' currency scale

-> Dense currency

Convert currency Discrete monetary value into a Dense monetary value.

denseFromDecimal

Arguments

:: DecimalConf
Config to use for parsing the decimal number.

Notice that a leading '-' or '+' will always be correctly interpreted, notwithstanding what the “leading '+'” policy is on the given DecimalConf.

-> Text
The raw string containing the decimal representation (e.g., "-1,234.56789").

-> Maybe (Dense currency)

Parses a decimal representation of a Dense.

denseToDecimal

Arguments

:: DecimalConf
Config to use for rendering the decimal number.

-> Approximation
Approximation to use if necessary in order to fit the Dense amount in as many decimal numbers as requested.

-> Dense currency
The monetary amount to render.

-> Text

Render a Dense monetary amount as a decimal number in a potentially lossy manner.

> denseToDecimal defaultDecimalConf Round
     (dense' (123456 % 100) :: Dense "USD")
"1234.56"

Discrete monetary values

type Discrete (currency :: Symbol) (unit :: Symbol) = Discrete' currency (UnitScale currency unit)

Discrete represents a discrete monetary value for a currency expresed as an integer amount of a particular unit. For example, with currency ~ "USD" and unit ~ "cent" you can represent United States Dollars to their full extent.

currency is usually a ISO-4217 currency code, but not necessarily.

Construct Discrete values using discrete, fromIntegral, fromInteger, discreteFromDense, discreteFromDecimal.

For example, if you want to represent GBP 21.05, where the smallest represetable unit for a GBP (United Kingdom Pound) is the penny, and 100 pennies equal 1 GBP (i.e., UnitScale "GBP" "penny" ~ '(100, 1)), then you can use:

discrete 2105 :: Discrete "GBP" "penny"

Because 2105 / 100 == 21.05.

data Discrete' (currency :: Symbol) (scale :: (Nat, Nat))

Discrete' represents a discrete monetary value for a currency expresed as amount of scale, which is a rational number expressed as (numerator, denominator).

You'll be using Discrete instead of Discrete' most of the time, which mentions the unit name (such as cent or centavo) instead of explicitely mentioning the unit scale.

discrete :: GoodScale scale => Integer -> Discrete' currency scale

Construct a Discrete value.

discreteCurrency

Arguments

:: (KnownSymbol currency, GoodScale scale)

=> Discrete' currency scale

-> Text

Discrete currency identifier.

> discreteCurrency (discrete 4 :: Discrete "USD" "cent")
"USD"

discreteFromDense

Arguments

:: forall currency scale. GoodScale scale

=> Approximation
Approximation to use if necessary in order to fit the Dense amount in the requested scale.

-> Dense currency

-> (Discrete' currency scale, Dense currency)

Approximate a Dense value x to the nearest value fully representable a given scale.

If the given Dense doesn't fit entirely in the scale, then a non-zero Dense reminder is returned alongside the Discrete approximation.

Proof that discreteFromDense doesn't lose money:

x == case discreteFromDense a x of
        (y, z) -> denseFromDiscrete y + z

discreteFromDecimal

Arguments

:: GoodScale scale

=> DecimalConf
Config to use for parsing the decimal number.

Notice that a leading '-' or '+' will always be correctly interpreted, notwithstanding what the “leading '+'” policy is on the given DecimalConf.

-> Text
The raw string containing the decimal representation (e.g., "-1,234.56789").

-> Maybe (Discrete' currency scale)

Parses a decimal representation of a Discrete.

Notice that parsing will fail unless the entire precision of the decimal number can be represented in the desired scale.

discreteToDecimal

Arguments

:: GoodScale scale

=> DecimalConf
Config to use for rendering the decimal number.

-> Approximation
Approximation to use if necessary in order to fit the Discrete amount in as many decimal numbers as requested.

-> Discrete' currency scale
The monetary amount to render.

-> Text

Render a Discrete' monetary amount as a decimal number in a potentially lossy manner.

This is simply a convenient wrapper around denseToDecimal:

discreteToDecimal ds a (dis :: Discrete' currency scale)
    == denseToDecimal ds a (denseFromDiscrete dis :: Dense currency)

In particular, the scale in Discrete' currency scale has no influence over the scale in which the decimal number is rendered. Change the scale with decimalConf_scale in order to modify that behavior.

Please refer to denseToDecimal for further documentation.

Currency scales

data Scale

This is the term-level representation of the “scale” we represent as (Nat, Nat) elsewhere in the type system (e.g., in GoodScale or UnitScale).

See UnitScale for a detailed description.

scaleFromRational :: Rational -> Maybe Scale

Construct a Scale from a positive, non-zero rational number.

scaleToRational :: Scale -> Rational

Obtain the Rational representation of a Scale.

scale

Arguments

:: forall proxy scale. GoodScale scale

=> proxy scale

-> Scale

Term-level representation of a currrency scale.

For example, the Scale for "USD" in "cent"s is 100/1. We can obtain a term-level representation for it using any of the following:

> scale (Proxy :: Proxy (UnitScale "USD" "cent"))
Scale (100 % 1)

> scale (Proxy :: CurrencyScale "USD")
Scale (100 % 1)

> scale (x :: Discrete "USD" "cent")
Scale (100 % 1)

The returned Rational is statically guaranteed to be a positive number.

type family UnitScale (currency :: Symbol) (unit :: Symbol) :: (Nat, Nat)

UnitScale currency unit is an rational number (expressed as '(numerator, denominator)) indicating how many pieces of unit fit in currency.

currency is usually a ISO-4217 currency code, but not necessarily.

The resulting (Nat, Nat), which is the type-level representation for what at the term-level we call Scale, will determine how to convert a Dense value into a Discrete value and vice-versa.

For example, there are 100 USD cents in 1 USD, so the scale for this relationship is:

type instance UnitScale "USD" "cent" = '(100, 1)

As another example, there is 1 dollar in USD, so the scale for this relationship is:

type instance UnitScale "USD" "dollar" = '(1, 1)

When using Discrete values to represent money, it will be impossible to represent an amount of currency smaller than unit. So, if you decide to use UnitScale "USD" "dollar" as your scale, you will not be able to represent values such as USD 3.50 or USD 21.87 becacuse they are not exact multiples of a dollar.

For some monetary values, such as precious metals, there is no smallest representable unit, since you can repeatedly split the precious metal many times before it stops being a precious metal. Nevertheless, for practical purposes we can make a sane arbitrary choice of smallest unit. For example, the base unit for XAU (Gold) is the troy ounce, which is too big to be considered the smallest unit, but we can arbitrarily choose the milligrain as our smallest unit, which is about as heavy as a single grain of table salt and should be sufficiently precise for all monetary practical purposes. A troy ounce equals 480000 milligrains.

type instance UnitScale "XAU" "milligrain" = '(480000, 1)

You can use other units such as milligrams for measuring XAU, for example. However, since the amount of milligrams in a troy ounce (31103.477) is not integral, we need to use rational with a denominator different than 1 to express it.

type instance UnitScale "XAU" "milligram" = '(31103477, 1000)

If you try to obtain the UnitScale of a currency without an obvious smallest representable unit, like XAU, you will get a compile error.

type family CurrencyScale (currency :: Symbol) :: (Nat, Nat)

If there exists a canonical smallest Scale that can fully represent the currency in all its denominations, then CurrencyScale currency will return such Scale. For example, CurrencyScale "USD" evaluates to UnitScale "USD" "cent".

type instance CurrencyScale "USD" = UnitScale "USD" "cent"

If the currency doesn't have a canonical smallest Scale, then CurrencyScale currency shall be left undefined or fail to compile with a TypeError. For example CurrencyScale "XAU" fails with ErrScaleNonCanonical "XAU".

type GoodScale (scale :: (Nat, Nat)) = (CmpNat 0 (Fst scale) ~ 'LT, CmpNat 0 (Snd scale) ~ 'LT, KnownNat (Fst scale), KnownNat (Snd scale))

Constraints to a scale (like the one returned by UnitScale currency unit) expected to always be satisfied. In particular, the scale is always guaranteed to be a positive rational number (infinity and notANumber are forbidden by GoodScale).

type family ErrScaleNonCanonical (currency :: Symbol) :: k where ...

A friendly TypeError to use for a currency that doesn't have a canonical small unit.

Equations

ErrScaleNonCanonical c = TypeError (('Text c :<>: 'Text " is not a currency with a canonical smallest unit,") :$$: 'Text "be explicit about the currency unit you want to use.")

Currency exchange

data ExchangeRate (src :: Symbol) (dst :: Symbol)

Exchange rate for converting monetary values of currency src into monetary values of currency dst by multiplying for it.

For example, if in order to convert USD to GBP we have to multiply by 1.2345, then we can represent this situaion using:

exchangeRate (12345 % 10000) :: Maybe (ExchangeRate "USD" "GBP")

exchangeRate :: Rational -> Maybe (ExchangeRate src dst)

Safely construct an ExchangeRate from a *positive* Rational number.

If the given Rational is non-positive, returns Nothing.

exchangeRate' :: Rational -> ExchangeRate src dst

Unsafely build an ExchageRate monetary value from a Rational value. Contrary to exchangeRate, this function *crashes* if the given Rational a value has zero as a denominator or when it is negative, with the former case being something very unlikely to happen unless the given Rational was itself unsafely constructed. Other than that, exchangeRate and exchangeRate' behave the same.

Prefer to use exchangeRate when dealing with Rational inputs from untrusted sources.

denominator x /= 0 && x > 0
  ⇒ exchangeRate x == Just (exchangeRate' x)

denominator x == 0 || x <= 0
  ⇒ undefined == exchangeRate' x

exchange :: ExchangeRate src dst -> Dense src -> Dense dst

Apply the ExchangeRate to the given Dense src monetary value.

Identity law:

exchange (exchangeRateRecip x) . exchange x  ==  id

Use the Identity law for reasoning about going back and forth between src and dst in order to manage any leftovers that might not be representable as a Discrete monetary value of src.

exchangeRateRecip :: ExchangeRate a b -> ExchangeRate b a

Reciprocal ExchangeRate.

This function retuns the reciprocal or multiplicative inverse of the given ExchangeRate, leading to the following identity law:

exchangeRateRecip . exchangeRateRecip   ==  id

Note: If ExchangeRate had a Fractional instance, then exchangeRateRecip would be the implementation of recip.

exchangeRateFromDecimal

Arguments

:: DecimalConf
Config to use for parsing the decimal number.

Notice that a leading '-' or '+' will always be correctly interpreted, notwithstanding what the “leading '+'” policy is on the given DecimalConf.

-> Text
The raw string containing the decimal representation (e.g., "1,234.56789").

-> Maybe (ExchangeRate src dst)

Parses a decimal representation of an ExchangeRate.

exchangeRateToDecimal

Arguments

:: DecimalConf
Config to use for rendering the decimal number.

-> Approximation
Approximation to use if necessary in order to fit the ExchangeRate amount in as many decimal numbers as requested.

-> ExchangeRate src dst
The ExchangeRate to render.

-> Text

Render a ExchangeRate as a decimal number in a potentially lossy manner.

> exchangeRateToDecimal defaultDecimalConf Round
      <$> (exchangeRate (123456 % 100) :: Maybe (ExchangeRate "USD" "EUR"))
Just "1,234.56"

exchangeRateToRational :: ExchangeRate src dst -> Rational

Obtain a Rational representation of the ExchangeRate.

This Rational is guaranteed to be a positive number.

Serializable representations

data SomeDense

A monomorphic representation of Dense that is easier to serialize and deserialize than Dense in case you don't know the type indexes involved.

If you are trying to construct a value of this type from some raw input, then you will need to use the mkSomeDense function.

In order to be able to effectively serialize a SomeDense value, you need to serialize the following three values (which are the eventual arguments to mkSomeDense):

someDenseCurrency

someDenseAmount

toSomeDense :: KnownSymbol currency => Dense currency -> SomeDense

Convert a Dense to a SomeDense for ease of serialization.

mkSomeDense

Arguments

:: Text
Currency. (someDenseCurrency)

-> Rational
Amount. (someDenseAmount)

-> Maybe SomeDense

Build a SomeDense from raw values.

This function is intended for deserialization purposes. You need to convert this SomeDense value to a Dense value in order to do any arithmetic operation on the monetary value.

fromSomeDense

Arguments

:: forall currency. KnownSymbol currency

=> SomeDense

-> Maybe (Dense currency)

Attempt to convert a SomeDense to a Dense, provided you know the target currency.

withSomeDense

Arguments

:: SomeDense

-> (forall currency. KnownSymbol currency => Dense currency -> r)

-> r

Convert a SomeDense to a Dense without knowing the target currency.

Notice that currency here can't leave its intended scope unless you can prove equality with some other type at the outer scope, but in that case you would be better off using fromSomeDense directly.

someDenseToDecimal

Arguments

:: DecimalConf
Config to use for rendering the decimal number.

-> Approximation
Approximation to use if necessary in order to fit the SomeDense amount in as many decimal numbers as requested.

-> SomeDense
The monetary amount to render.

-> Text

Like denseToDecimal, but takes a SomeDense as input.

someDenseCurrency :: SomeDense -> Text

Currency name.

someDenseAmount :: SomeDense -> Rational

Currency unit amount.

data SomeDiscrete

A monomorphic representation of Discrete that is easier to serialize and deserialize than Discrete in case you don't know the type indexes involved.

If you are trying to construct a value of this type from some raw input, then you will need to use the mkSomeDiscrete function.

In order to be able to effectively serialize a SomeDiscrete value, you need to serialize the following four values (which are the eventual arguments to mkSomeDiscrete):

someDiscreteCurrency

someDiscreteScale

someDiscreteAmount

toSomeDiscrete

Arguments

:: (KnownSymbol currency, GoodScale scale)

=> Discrete' currency scale

-> SomeDiscrete

Convert a Discrete to a SomeDiscrete for ease of serialization.

mkSomeDiscrete

Arguments

:: Text
Currency name. (someDiscreteCurrency)

-> Scale
Scale. Positive, non-zero. (someDiscreteScale)

-> Integer
Amount of unit. (someDiscreteAmount)

-> SomeDiscrete

Build a SomeDiscrete from raw values.

This function is intended for deserialization purposes. You need to convert this SomeDiscrete value to a Discrete vallue in order to do any arithmetic operation on the monetary value.

fromSomeDiscrete

Arguments

:: forall currency scale. (KnownSymbol currency, GoodScale scale)

=> SomeDiscrete

-> Maybe (Discrete' currency scale)

Attempt to convert a SomeDiscrete to a Discrete, provided you know the target currency and unit.

withSomeDiscrete

Arguments

:: forall r. SomeDiscrete

-> (forall currency scale. (KnownSymbol currency, GoodScale scale) => Discrete' currency scale -> r)

-> r

Convert a SomeDiscrete to a Discrete without knowing the target currency and unit.

Notice that currency and unit here can't leave its intended scope unless you can prove equality with some other type at the outer scope, but in that case you would be better off using fromSomeDiscrete directly.

Notice that you may need to add an explicit type to the result of this function in order to keep the compiler happy.

someDiscreteToDecimal

Arguments

:: DecimalConf
Config to use for rendering the decimal number.

-> Approximation
Approximation to use if necessary in order to fit the SomeDiscrete amount in as many decimal numbers as requested.

-> SomeDiscrete
The monetary amount to render.

-> Text

Like discreteToDecimal, but takes a SomeDiscrete as input.

someDiscreteCurrency :: SomeDiscrete -> Text

Currency name.

someDiscreteScale :: SomeDiscrete -> Scale

Positive, non-zero.

someDiscreteAmount :: SomeDiscrete -> Integer

Amount of currency unit.

data SomeExchangeRate

A monomorphic representation of ExchangeRate that is easier to serialize and deserialize than ExchangeRate in case you don't know the type indexes involved.

If you are trying to construct a value of this type from some raw input, then you will need to use the mkSomeExchangeRate function.

In order to be able to effectively serialize an SomeExchangeRate value, you need to serialize the following four values (which are the eventual arguments to mkSomeExchangeRate):

someExchangeRateSrcCurrency

someExchangeRateDstCurrency

someExchangeRateRate

toSomeExchangeRate

Arguments

:: (KnownSymbol src, KnownSymbol dst)

=> ExchangeRate src dst

-> SomeExchangeRate

Convert a ExchangeRate to a SomeDiscrete for ease of serialization.

mkSomeExchangeRate

Arguments

:: Text
Source currency name. (someExchangeRateSrcCurrency)

-> Text
Destination currency name. (someExchangeRateDstCurrency)

-> Rational
Exchange rate . Positive, non-zero. (someExchangeRateRate)

-> Maybe SomeExchangeRate

Build a SomeExchangeRate from raw values.

This function is intended for deserialization purposes. You need to convert this SomeExchangeRate value to a ExchangeRate value in order to do any arithmetic operation with the exchange rate.

fromSomeExchangeRate

Arguments

:: forall src dst. (KnownSymbol src, KnownSymbol dst)

=> SomeExchangeRate

-> Maybe (ExchangeRate src dst)

Attempt to convert a SomeExchangeRate to a ExchangeRate, provided you know the target src and dst types.

withSomeExchangeRate

Arguments

:: SomeExchangeRate

-> (forall src dst. (KnownSymbol src, KnownSymbol dst) => ExchangeRate src dst -> r)

-> r

Convert a SomeExchangeRate to a ExchangeRate without knowing the target currency and unit.

Notice that src and dst here can't leave its intended scope unless you can prove equality with some other type at the outer scope, but in that case you would be better off using fromSomeExchangeRate directly.

someExchangeRateToDecimal

Arguments

:: DecimalConf
Config to use for rendering the decimal number.

-> Approximation
Approximation to use if necessary in order to fit the SomeExchangeRate amount in as many decimal numbers as requested.

-> SomeExchangeRate
The SomeExchangeRate to render.

-> Text

Like exchangeRateToDecimal, but takes a SomeExchangeRate as input.

someExchangeRateSrcCurrency :: SomeExchangeRate -> Text

Source currency name.

someExchangeRateDstCurrency :: SomeExchangeRate -> Text

Destination currency name.

someExchangeRateRate :: SomeExchangeRate -> Rational

Exchange rate. Positive, non-zero.

Miscellaneous

data Approximation

Method for approximating a fractional number to an integer number.

Constructors

Round
Approximate x to the nearest integer, or to the nearest even integer if x is equidistant between two integers.

Floor
Approximate x to the nearest integer less than or equal to x.

Ceiling
Approximate x to the nearest integer greater than or equal to x.

Truncate
Approximate x to the nearest integer betwen 0 and x, inclusive.

HalfEven
Approximate x to the nearest even integer, when equidistant from the nearest two integers. This is also known as “Bankers Rounding”.

HalfAwayFromZero
Approximate x to the nearest integer, or to the nearest integer away from zero if x is equidistant between to integers. This is known as “kaufmännisches Runden” in German speaking countries.

Decimal config

data DecimalConf

Config to use when rendering or parsing decimal numbers.

See defaultDecimalConf.

Constructors

DecimalConf

Fields

decimalConf_separators :: !Separators

Decimal and thousands separators to use when rendering the decimal number. Construct one with mkSeparators, or pick a ready made one like separatorsDot or separatorsDotNarrownbsp.

decimalConf_leadingPlus :: !Bool

Whether to render a leading '+' sign in case the amount is positive.

decimalConf_digits :: !Word8

Number of decimal numbers to render, if any.

decimalConf_scale :: !Scale

Scale used to when rendering the decimal number. This is useful if, for example, you want to render a “number of cents” rather than a “number of dollars” as the whole part of the decimal number when rendering a USD amount. It's particularly useful when rendering currencies such as XAU, where one might prefer to render amounts as a number of grams, rather than as a number of troy-ounces.

Set this to 1 if you don't care.

For example, when rendering render dense' (123 % 100) :: Dense "USD" as a decimal number with two decimal places, a scale of 1 (analogous to UnitScale "USD" "dollar") would render 1 as the integer part and 23 as the fractional part, whereas a scale of 100 (analogous UnitScale "USD" "cent") would render 123 as the integer part and 00 as the fractional part.

You can easily obtain the scale for a particular currency and unit combination using the scale function.

Important: Generally, you will want this number to be 1 or larger. This is because scales in the range (0, 1) can be particularly lossy unless the number of decimal digits is sufficiently large.

defaultDecimalConf :: DecimalConf

Default DecimalConf.

No leading '+' sign

No thousands separator

Decimal separator is '.'

2 decimal digits

A scale of 1

That is, something like 1.23 or -1234567.89.

Separators

data Separators

Decimal and thousands separators used when rendering or parsing a decimal number.

Use mkSeparators to construct.

mkSeparators

Arguments

:: Char
Decimal separator (i.e., the '.' in 1,234.56789)

-> Maybe Char
Thousands separator for the integer part, if any (i.e., the ',' in 1,234.56789).

-> Maybe Separators

Construct Separators to use with in DecimalConf.

The separators can't be an ASCII digit nor control character, and they must be different from each other.

separatorsComma :: Separators

1234567,89

separatorsCommaDot :: Separators

1.234.567,89

separatorsCommaNarrownbsp :: Separators

1 234 567,89

The whitespace is Unicode's NARROW NO-BREAK SPACE (U+202f, 8239, '8239').

separatorsCommaNbsp :: Separators

1 234 567,89

The whitespace is Unicode's NO-BREAK SPACE (U+00a0, 160, '160').

separatorsCommaThinsp :: Separators

1 234 567,89

The whitespace is Unicode's THIN SPACE (U+2009, 8201, '8201').

separatorsCommaSpace :: Separators

1 234 567,89

The whitespace is ASCII's SPC (U+0020, 32, '32').

separatorsDot :: Separators

1234567.89

separatorsDotComma :: Separators

1,234,567.89

separatorsDotNarrownbsp :: Separators

1 234 567.89

The whitespace is Unicode's NARROW NO-BREAK SPACE (U+202f, 8239, '8239').

separatorsDotThinsp :: Separators

1 234 567.89

The whitespace is Unicode's THIN SPACE (U+2009, 8201, '8201').

separatorsDotNbsp :: Separators

1 234 567.89

The whitespace is Unicode's NO-BREAK SPACE (U+00a0, 160, '160').

separatorsDotSpace :: Separators

1 234 567.89

The whitespace is ASCII's SPACE (U+0020, 32, '32').